type hierarchy for AslVisualizationCDK
